feat: move and adjust layout components
move header wrapper and nav components to layout components
In the Nav.vue file, updates have been made to the navigation bar. This includes changes to the dropdown menus, link behaviors, and responsive design. The aim is to provide a smoother navigation experience across different device sizes.
